[Hounslow] Add general enquiries functionality
This functionality allows a cobrand to replace the /contact form
with a form that creates hidden reports which are sent via Open311.

The form also allows file uploads in addition to photos.

This functionality is currently enabled for the Hounslow cobrand and
others cobrands can enable it by defining setup_general_enquiries_stash
which primarily sets up the appropriate categories and default values
for the report.

+++ perllib/FixMyStreet/App/Controller/Contact.pm
@@ -26,9 +26,14 @@ Functions to run on both GET and POST contact requests.
 
 sub auto : Private {
     my ($self, $c) = @_;
+    $c->forward('/auth/get_csrf_token');
+}
+
+sub begin : Private {
+    my ($self, $c) = @_;
+    $c->forward('/begin');
     $c->forward('setup_request');
     $c->forward('determine_contact_type');
-    $c->forward('/auth/get_csrf_token');
 }
 
 =head2 index
@@ -106,7 +111,14 @@ sub determine_contact_type : Private {
             $c->stash->{rejecting_report} = 1;
         }
     } elsif ( $c->cobrand->abuse_reports_only ) {
-        $c->detach( '/page_error_404_not_found' );
+        # General enquiries replaces contact form if enabled
+        if ( $c->cobrand->can('setup_general_enquiries_stash') ) {
+            $c->res->redirect( '/contact/enquiry' );
+            $c->detach;
+            return 1;
+        } else {
+            $c->detach( '/page_error_404_not_found' );
+        }
     }
 
     return 1;
